+++ lib/dexter/sources/statement_source.rb
@@ -1,11 +1,13 @@
 module Dexter
   class StatementSource
-    def initialize(statement)
-      @statement = statement
+    def initialize(statements)
+      @statements = statements
     end
 
     def perform(collector)
-      collector.add(@statement, 0, 0, true)
+      @statements.each do |statement|
+        collector.add(statement, 0, 0, true)
+      end
     end
   end
 end